Refactor HttpControlService
